# api/contract/openapi.yaml
openapi: 3.0.0
paths:
...
  /{sessionId}/contacts:
    get:
      operationId: ContactController_getContacts
      parameters:
        - name: offset
          required: false
          in: query
          description: 'Offset (default: 0)'
          schema:
            type: number
        - name: limit
          required: false
          in: query
          description: 'Limit (default: 50)'
          schema:
            type: number
        - name: search
          required: false
          in: query
          description: Search by name or JID
          schema:
            type: string
        - name: sessionId
          required: true
          in: path
          schema: {}
      responses:
        '200':
          description: ''
      security:
        - x-api-key: []
      summary: Get all contacts (from database)
      tags:
        - Contact
  /{sessionId}/contacts/check:
    post:
      operationId: ContactController_checkNumber
      parameters:
        - name: sessionId
          required: true
          in: path
          schema: {}
      responses:
        '201':
          description: ''
      security:
        - x-api-key: []
      summary: Check if numbers exist on WhatsApp
      tags:
        - Contact
  /{sessionId}/contacts/{jid}/profile-picture:
    get:
      operationId: ContactController_getProfilePicture
      parameters:
        - name: highRes
          required: false
          in: query
          schema: {}
        - name: jid
          required: true
          in: path
          schema: {}
        - name: sessionId
          required: true
          in: path
          schema: {}
      responses:
        '200':
          description: ''
      security:
        - x-api-key: []
      summary: Get profile picture URL
      tags:
        - Contact
  /{sessionId}/contacts/{jid}/business-profile:
    get:
      operationId: ContactController_getBusinessProfile
      parameters:
        - name: jid
          required: true
          in: path
          schema: {}
        - name: sessionId
          required: true
          in: path
          schema: {}
      responses:
        '200':
          description: ''
      security:
        - x-api-key: []
      summary: Get business profile
      tags:
        - Contact
  /{sessionId}/contacts/profile/business-profile:
    post:
      operationId: ContactController_updateBusinessProfile
      parameters:
        - name: sessionId
          required: true
          in: path
          schema: {}
      responses:
        '201':
          description: ''
      security:
        - x-api-key: []
      summary: Update own business profile (requires business account)
      tags:
        - Contact
  /{sessionId}/contacts/{jid}/status:
    get:
      operationId: ContactController_getStatus
      parameters:
        - name: jid
          required: true
          in: path
          schema: {}
        - name: sessionId
          required: true
          in: path
          schema: {}
      responses:
        '200':
          description: ''
      security:
        - x-api-key: []
      summary: Get contact about/status
      tags:
        - Contact
  /{sessionId}/contacts/{jid}/block:
    post:
      operationId: ContactController_block
      parameters:
        - name: jid
          required: true
          in: path
          schema: {}
        - name: sessionId
          required: true
          in: path
          schema: {}
      responses:
        '201':
          description: ''
      security:
        - x-api-key: []
      summary: Block a contact
      tags:
        - Contact
  /{sessionId}/contacts/{jid}/unblock:
    post:
      operationId: ContactController_unblock
      parameters:
        - name: jid
          required: true
          in: path
          schema: {}
        - name: sessionId
          required: true
          in: path
          schema: {}
      responses:
        '201':
          description: ''
      security:
        - x-api-key: []
      summary: Unblock a contact
      tags:
        - Contact
  /{sessionId}/contacts/profile/picture:
    post:
      operationId: ContactController_updateProfilePicture
      parameters:
        - name: sessionId
          required: true
          in: path
          schema: {}
      responses:
        '201':
          description: ''
      security:
        - x-api-key: []
      summary: Update own profile picture
      tags:
        - Contact
  /{sessionId}/contacts/profile/name:
    post:
      operationId: ContactController_updateProfileName
      parameters:
        - name: sessionId
          required: true
          in: path
          schema: {}
      responses:
        '201':
          description: ''
      security:
        - x-api-key: []
      summary: Update own profile name
      tags:
        - Contact
  /{sessionId}/contacts/profile/status:
    post:
      operationId: ContactController_updateProfileStatus
      parameters:
        - name: sessionId
          required: true
          in: path
          schema: {}
      responses:
        '201':
          description: ''
      security:
        - x-api-key: []
      summary: Update own about/status text
      tags:
        - Contact
...
info:
  title: Baileys WhatsApp API
  description: Comprehensive REST API for WhatsApp using Baileys library.
  version: 1.0.0
  contact: {}
...
servers: []
components:
  securitySchemes:
    x-api-key:
      type: apiKey
      in: header
      name: x-api-key
...